Refactor: split sidebar component and move file
