+++ src/app/(route)/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 
 import localFont from 'next/font/local';
 
-import { Sidebar } from '@/components/common/Sidebar';
+import { Sidebar } from '@/components/Sidebar';
 import QueryProvider from '@/provider/QueryProvider';
 import '@/styles/globals.css';
